+++ src/main/java/com/lec/spring/repository/ProviderRepository.java
@@ -9,13 +9,13 @@ import java.util.List;
 @Mapper
 public interface ProviderRepository {
     List<ProvLodging> findLodgings(Long userId);
-    ProvLodging findAllDetails(int lodgingId);
+    ProvLodging findAllDetails(Long lodgingId);
     void saveLodging(ProvLodging lodging);
 
     List<ProvLodging> findByUserId(Long userId);
 
     void updateLodging(ProvLodging lodging);
 
-    void deleteLodging(int lodgingId);
+    void deleteLodging(Long lodgingId);
 
 }
