+++ src/main/java/com/lec/spring/service/ProviderServiceImpl.java
@@ -27,7 +27,7 @@ public class ProviderServiceImpl implements ProviderService {
     }
 
     @Override
-    public ProvLodging getAllDetails(int lodgingId) {
+    public ProvLodging getAllDetails(Long lodgingId) {
         return providerRepository.findAllDetails(lodgingId);
     }
 
@@ -52,7 +52,7 @@ public class ProviderServiceImpl implements ProviderService {
     }
 
     @Override
-    public void deleteLodging(int lodgingId) {
+    public void deleteLodging(Long lodgingId) {
         providerRepository.deleteLodging(lodgingId);
     }
 }
